creating more separation of forward and reverse mode
